+++ RELEASES.md
@@ -6,6 +6,7 @@
 - Register types in the reflection system.
 - added support in `ActionDiff` for value and axis_pair changes
 - Added `InputMap::Clear`.
+- Fixed [a bug](https://github.com/Leafwing-Studios/leafwing-input-manager/issues/430) related to incorrect axis data in `Chord` when not all buttons are pressed.
 
 ## Version 0.11.2